#  Significantly improve encoding performance