## Benchmark

### BEFORE

```
goos: darwin
goarch: amd64
pkg: benchmark
Benchmark_Encode_SmallStruct_EncodingJson-16                     2142819               552 ns/op             256 B/op          2 allocs/op
Benchmark_Encode_SmallStruct_JsonIter-16                         2089278               582 ns/op             264 B/op          3 allocs/op
Benchmark_Encode_SmallStruct_GoJay-16                            2951632               417 ns/op             624 B/op          2 allocs/op
Benchmark_Encode_SmallStruct_SegmentioJson-16                    4154019               285 ns/op             256 B/op          2 allocs/op
Benchmark_Encode_SmallStruct_GoJson-16                           4204060               289 ns/op             144 B/op          1 allocs/op
Benchmark_Encode_SmallStructCached_EncodingJson-16               2392526               518 ns/op             144 B/op          1 allocs/op
Benchmark_Encode_SmallStructCached_JsonIter-16                   2060712               534 ns/op             152 B/op          2 allocs/op
Benchmark_Encode_SmallStructCached_GoJay-16                      3347031               360 ns/op             512 B/op          1 allocs/op
Benchmark_Encode_SmallStructCached_SegmentioJson-16              5050892               245 ns/op             144 B/op          1 allocs/op
Benchmark_Encode_SmallStructCached_GoJson-16                     4067506               286 ns/op             144 B/op          1 allocs/op
Benchmark_Encode_MediumStruct_EncodingJson-16                     761324              1696 ns/op             632 B/op         15 allocs/op
Benchmark_Encode_MediumStruct_JsonIter-16                         735626              1546 ns/op             640 B/op         16 allocs/op
Benchmark_Encode_MediumStruct_GoJay-16                           1000000              1107 ns/op             824 B/op         15 allocs/op
Benchmark_Encode_MediumStruct_SegmentioJson-16                    961773              1078 ns/op             633 B/op         15 allocs/op
Benchmark_Encode_MediumStruct_GoJson-16                          2143916               552 ns/op             320 B/op          1 allocs/op
Benchmark_Encode_MediumStructCached_EncodingJson-16              1000000              1129 ns/op             320 B/op          1 allocs/op
Benchmark_Encode_MediumStructCached_JsonIter-16                  1000000              1012 ns/op             328 B/op          2 allocs/op
Benchmark_Encode_MediumStructCached_GoJay-16                     1725355               725 ns/op             512 B/op          1 allocs/op
Benchmark_Encode_MediumStructCached_SegmentioJson-16             1812170               644 ns/op             320 B/op          1 allocs/op
Benchmark_Encode_MediumStructCached_GoJson-16                    2077965               536 ns/op             320 B/op          1 allocs/op
Benchmark_Encode_LargeStruct_EncodingJson-16                       40851             31579 ns/op           14665 B/op        319 allocs/op
Benchmark_Encode_LargeStruct_JsonIter-16                           41943             28746 ns/op           14671 B/op        320 allocs/op
Benchmark_Encode_LargeStruct_GoJay-16                              46891             25085 ns/op           28051 B/op        323 allocs/op
Benchmark_Encode_LargeStruct_SegmentioJson-16                      57402             20780 ns/op           14647 B/op        319 allocs/op
Benchmark_Encode_LargeStruct_GoJson-16                             59654             20083 ns/op           14669 B/op        319 allocs/op
Benchmark_Encode_LargeStructCached_EncodingJson-16                 71298             16655 ns/op            4867 B/op          1 allocs/op
Benchmark_Encode_LargeStructCached_JsonIter-16                     79042             15570 ns/op            4875 B/op          2 allocs/op
Benchmark_Encode_LargeStructCached_GoJay-16                        89496             12677 ns/op           18314 B/op          5 allocs/op
Benchmark_Encode_LargeStructCached_SegmentioJson-16               132954              9213 ns/op            4867 B/op          1 allocs/op
Benchmark_Encode_LargeStructCached_GoJson-16                      149628              8324 ns/op            4870 B/op          1 allocs/op
PASS
ok      benchmark       44.499s
```

### AFTER

```
goos: darwin
goarch: amd64
pkg: benchmark
Benchmark_Encode_SmallStruct_EncodingJson-16                     2135508               549 ns/op             256 B/op          2 allocs/op
Benchmark_Encode_SmallStruct_JsonIter-16                         1990358               604 ns/op             264 B/op          3 allocs/op
Benchmark_Encode_SmallStruct_GoJay-16                            2954782               404 ns/op             624 B/op          2 allocs/op
Benchmark_Encode_SmallStruct_SegmentioJson-16                    4164482               284 ns/op             256 B/op          2 allocs/op
Benchmark_Encode_SmallStruct_GoJson-16                           5158462               235 ns/op             144 B/op          1 allocs/op
Benchmark_Encode_SmallStructCached_EncodingJson-16               2426574               493 ns/op             144 B/op          1 allocs/op
Benchmark_Encode_SmallStructCached_JsonIter-16                   2209437               551 ns/op             152 B/op          2 allocs/op
Benchmark_Encode_SmallStructCached_GoJay-16                      3399403               356 ns/op             512 B/op          1 allocs/op
Benchmark_Encode_SmallStructCached_SegmentioJson-16              5200904               229 ns/op             144 B/op          1 allocs/op
Benchmark_Encode_SmallStructCached_GoJson-16                     5208650               228 ns/op             144 B/op          1 allocs/op
Benchmark_Encode_MediumStruct_EncodingJson-16                     778935              1561 ns/op             632 B/op         15 allocs/op
Benchmark_Encode_MediumStruct_JsonIter-16                         754591              1471 ns/op             640 B/op         16 allocs/op
Benchmark_Encode_MediumStruct_GoJay-16                           1000000              1054 ns/op             824 B/op         15 allocs/op
Benchmark_Encode_MediumStruct_SegmentioJson-16                   1202144              1005 ns/op             633 B/op         15 allocs/op
Benchmark_Encode_MediumStruct_GoJson-16                          2439882               498 ns/op             320 B/op          1 allocs/op
Benchmark_Encode_MediumStructCached_EncodingJson-16              1000000              1124 ns/op             320 B/op          1 allocs/op
Benchmark_Encode_MediumStructCached_JsonIter-16                  1000000              1030 ns/op             328 B/op          2 allocs/op
Benchmark_Encode_MediumStructCached_GoJay-16                     1757950               681 ns/op             512 B/op          1 allocs/op
Benchmark_Encode_MediumStructCached_SegmentioJson-16             1875118               631 ns/op             320 B/op          1 allocs/op
Benchmark_Encode_MediumStructCached_GoJson-16                    2490054               484 ns/op             320 B/op          1 allocs/op
Benchmark_Encode_LargeStruct_EncodingJson-16                       42511             27779 ns/op           14670 B/op        319 allocs/op
Benchmark_Encode_LargeStruct_JsonIter-16                           43094             26771 ns/op           14677 B/op        320 allocs/op
Benchmark_Encode_LargeStruct_GoJay-16                              49084             24243 ns/op           28051 B/op        323 allocs/op
Benchmark_Encode_LargeStruct_SegmentioJson-16                      60148             19914 ns/op           14647 B/op        319 allocs/op
Benchmark_Encode_LargeStruct_GoJson-16                             64729             18682 ns/op           14674 B/op        319 allocs/op
Benchmark_Encode_LargeStructCached_EncodingJson-16                 70771             16412 ns/op            4867 B/op          1 allocs/op
Benchmark_Encode_LargeStructCached_JsonIter-16                     76849             15427 ns/op            4875 B/op          2 allocs/op
Benchmark_Encode_LargeStructCached_GoJay-16                        92294             12672 ns/op           18314 B/op          5 allocs/op
Benchmark_Encode_LargeStructCached_SegmentioJson-16               131605              8735 ns/op            4867 B/op          1 allocs/op
Benchmark_Encode_LargeStructCached_GoJson-16                      162507              7230 ns/op            4867 B/op          1 allocs/op
PASS
ok      benchmark       44.672s
```

## Changed Summary

- Use local buffer to append new buffer 
- Replace (unsigned)integer encoding with `github.com/segmentio/encoding/json` logic
- Replace string encoding with `github.com/segmentio/encoding/json` logic
- Remove unnecessary sync.Pool